+ Adding the diagram generator python code. + Adding Evaluator + Improving arc primitive implementation.
! Frustum culling needs to be added.
! Using one big buffer vs. multiple small buffers need to be compared.
! Diagrams need to be generated.

+++ glDynamicTessellation_CPU/src/Application.h
@@ -106,7 +106,7 @@ public:
     GLuint      fbo_textures[3];
     GLuint      quad_vao;
 
-    Evaluator<int64_t, int64_t> trianglesPerSecondEvaluator;
+    Evaluator<int64_t, float> trianglesPerSecondEvaluator;
     
 
 	std::vector<ArcSegment>  arcSegments;
